# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Mazayada API Documentation'
  description: 'Mazayada mobile API (v1). Token-authenticated JSON API for the Flutter client covering auctions and the citizen dashboard. All responses use a unified envelope ({data, message, meta} on success; {message, errors} on error). Money is returned in dinars; send Accept-Language (ar|fr|en) to localize responses.'
  version: 1.0.0
servers:
  -
    url: 'https://mazayada.findosystem.com'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtain an <b>access token</b> from <code>POST /api/v1/auth/login</code> or <code>POST /api/v1/auth/verify-otp</code> and send it as <code>Authorization: Bearer &lt;access_token&gt;</code>. Access tokens are short-lived — use the <b>refresh token</b> at <code>POST /api/v1/auth/refresh</code> (send the refresh token as the bearer) to obtain a new pair.'
security:
  -
    default: []
paths:
...
  /api/v1/my-auctions:
    get:
      summary: 'My auctions'
      operationId: myAuctions
      description: "The user's participations grouped by tab (active | won | lost | upcoming)."
      parameters:
        -
          in: query
          name: tab
          description: 'One of active, won, lost, upcoming (default active).'
          example: active
          required: false
          schema:
            type: string
            description: 'One of active, won, lost, upcoming (default active).'
            example: active
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'يلزم تسجيل الدخول.'
                properties:
                  message:
                    type: string
                    example: 'يلزم تسجيل الدخول.'
      tags:
        - Dashboard
      security: []
...
  /api/v1/kyc:
    get:
      summary: 'KYC status'
      operationId: kYCStatus
      description: "The user's KYC state, which documents are on file, and whether a (re)submit\nis currently allowed."
      parameters:
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'يلزم تسجيل الدخول.'
                properties:
                  message:
                    type: string
                    example: 'يلزم تسجيل الدخول.'
      tags:
        - KYC
      security: []
...
  '/api/v1/payments/{ref}/status':
    get:
      summary: 'Payment status'
      operationId: paymentStatus
      description: "Returns the status of every payment row sharing a gateway reference for the\nauthenticated user. Use this to poll after the gateway web view returns."
      parameters:
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'يلزم تسجيل الدخول.'
                properties:
                  message:
                    type: string
                    example: 'يلزم تسجيل الدخول.'
      tags:
        - Payments
      security: []
    parameters:
      -
        in: path
        name: ref
        description: ''
        example: a1f6e506-27d4-496f-85a5-03541c909118
        required: true
        schema:
          type: string
...